requirement.txt
# Core Python
numpy
pandas

# Visualization
matplotlib
seaborn

# Audio Processing
librosa
soundfile

# Machine Learning
scikit-learn
xgboost

# Speech-to-Text (ASR)
torch
openai-whisper
openai>=1.0.0
python-dotenv

# Grammar Analysis
language-tool-python

# Web Application
flask

# Utilities
tqdm
joblib


flask
werkzeug
scipy
librosa
soundfile



